requirements.txt
aniso8601==9.0.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
psycopg2-binary==2.8.6
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
Werkzeug==1.0.1
